+++ spec/scanner/escapes_spec.rb
@@ -26,17 +26,31 @@ RSpec.describe('Escape scanning') do
     'a\u{10FFFF}c'    => [1, :escape,  :codepoint_list,   '\u{10FFFF}',     1,  11],
 
     /a\cBc/           => [1, :escape,  :control,          '\cB',            1,  4],
+    /a\c^c/           => [1, :escape,  :control,          '\c^',            1,  4],
+    /a\c\n/           => [1, :escape,  :control,          '\c\n',           1,  5],
+    /a\c\\b/          => [1, :escape,  :control,          '\c\\\\',         1,  5],
     /a\C-bc/          => [1, :escape,  :control,          '\C-b',           1,  5],
+    /a\C-^b/          => [1, :escape,  :control,          '\C-^',           1,  5],
+    /a\C-\nb/         => [1, :escape,  :control,          '\C-\n',          1,  6],
+    /a\C-\\b/         => [1, :escape,  :control,          '\C-\\\\',        1,  6],
     /a\c\M-Bc/n       => [1, :escape,  :control,          '\c\M-B',         1,  7],
     /a\C-\M-Bc/n      => [1, :escape,  :control,          '\C-\M-B',        1,  8],
 
     /a\M-Bc/n         => [1, :escape,  :meta_sequence,    '\M-B',           1,  5],
-    /a\M-\C-Bc/n      => [1, :escape,  :meta_sequence,    '\M-\C-B',        1,  8],
     /a\M-\cBc/n       => [1, :escape,  :meta_sequence,    '\M-\cB',         1,  7],
+    /a\M-\c^/n        => [1, :escape,  :meta_sequence,    '\M-\c^',         1,  7],
+    /a\M-\c\n/n       => [1, :escape,  :meta_sequence,    '\M-\c\n',        1,  8],
+    /a\M-\c\\/n       => [1, :escape,  :meta_sequence,    '\M-\c\\\\',      1,  8],
+    /a\M-\C-Bc/n      => [1, :escape,  :meta_sequence,    '\M-\C-B',        1,  8],
+    /a\M-\C-\\/n      => [1, :escape,  :meta_sequence,    '\M-\C-\\\\',     1,  9],
 
     'ab\\\xcd'        => [1, :escape,  :backslash,        '\\\\',           2,  4],
     'ab\\\0cd'        => [1, :escape,  :backslash,        '\\\\',           2,  4],
     'ab\\\Kcd'        => [1, :escape,  :backslash,        '\\\\',           2,  4],
+
+    'ab\^cd'          => [1, :escape,  :bol,              '\^',             2,  4],
+    'ab\$cd'          => [1, :escape,  :eol,              '\$',             2,  4],
+    'ab\[cd'          => [1, :escape,  :set_open,         '\[',             2,  4],
   }
 
   tests.each_with_index do |(pattern, (index, type, token, text, ts, te)), count|